refactor(workflows): extract gate routes + helpers, delete populateWorkflowRoutes (C2c)
Phase C2c of the workflows/index.ts split — the final route group. Pulls the
human-gate routes and their helpers out of the module-load
populateWorkflowRoutes mutation, which is now deleted entirely.

- lib/gate-audit.ts     — getGateDescription + buildGateAuditPayload, deduped
                          from the module/route/activate triple-scope copies.
- lib/gate-html.ts      — formValue / gateDecisionHtmlResponse / escapeHtml,
                          the durable server-rendered approval fallback page.
- lib/gate-settings.ts  — the gateNotificationSettings seed/fallback accessor
                          (setGateSettings / getGateSettings / activeGateSettings).
                          activeGateSettings still prefers the live notifications
                          store; onSettingsChange still updates the store only —
                          behavior identical, the former module global is gone.
- lib/routes/gates.ts   — exports `gateRoutes`: approve, reject, the decision
                          page GET + POST, /gates/pending, /gates/status, plus
                          the webApprover helper.

populateWorkflowRoutes is DELETED; definePlugin now spreads
`[...definitionRoutes, ...instanceRoutes, ...gateRoutes]`. Every workflow route
now lives in a lib/routes/ module. index.ts sheds ~9 now-unused imports and the
bottom HTML helpers.

Pure behavior-preserving relocation — route objects byte-identical, the
gate-settings accessor preserves the exact store-then-seed resolution order.

Verified: typecheck + eslint + workflows suite (420/0) + full suite (5124/0)
+ binary build + isolated-home boot smoke (GET /gates/pending + /gates/status
→ 200; /gates/x/decision missing params + POST /gates/x/approve bad body → 400;
/instances + /definitions still 200).
